fixed errors and clean code (#38)
* delete kotlin

* define API-based methods for blockchains

* delete method

* put method

* APi-based methods of ledger account

* fixed API url

* remove explicitly importing

* change class to LedgerAccount

* 20 seconds timeout

* change ledger account name

* fix bcash

* ledger customer

* order book

* add Record

* ledger Subscription

* ledger Transaction

* lombok compile

* API-bases methods for KMS

* lombok compile

* applied lombok

* add details for blockchain API, test code

* update ApiKey and BaseUrl.java

* update API details and test code

* update code

* update details for API

* update details for API

* add details to APIs and test code

* remove unused import package

* Ledger account test

* test create account for the ledger

* test create account for the ledger

* unit test ledger account

* unit test customer ledger

* unit test customer ledger

* unit test order book

* unit test

* unit test

* clean code

* validate objects

* validate objects

* validate CreateAccount

* validate CreateAccount

* validate object and unit test

* clean code

* transaction for bitcoin and litecoin

* apply async

* apply async

* change names

* sign KMS transaction

* bcach transaction

* fix litecoin

* fix errors

* xrp transaction

* xrp with kms transaction

* send xrp transaction

* xrp wallet

* btc wallet

* change test package

* bip32

* derivation path

* refactor code

* refactor code

* fix input parameter

* don not use singleton objects to prevent from risk of race condition

* clean code

* generate private keys and addresses

* ethe wallet

* ethe wallet

* bch, vet wallets

* eth private key

* eth transaction

* create specific networks

* eth transaction

* eth transaction

* eth transaction

* eth transaction test

* Xlm transaction and eth transaction test

* offchain btc

* clean code

* eth offchain

* eth offchain

* ltc offchain

* xlm offchain

* update README.md

* update README.md

* generate Javadocs

* generate Javadocs

* xrp offchain

* fixed errors

* fixed errors

* apidocs

* kms offchain

* update Javadocs

* update Javadocs

* update Javadocs

* delete apidocs

* add java docs

* update encode and decode transaaction

* update with bitcoincashj

* bch offchain

* bcash address

* bcash offchain

* vet Transaction

* review code

* review code

* log message

* update logic as tatumjs

* bcash transaction

* update

* update variables

* testing

* update variables

* fixed errors

* fixed errors

* update document

* update variables

* refactor

* refactor

* refactor

* bech32 for change address

* refactor

+++ src/main/java/io/tatum/model/response/bch/BchTxVin.java
@@ -17,9 +17,9 @@ import java.math.BigDecimal;
 public class BchTxVin {
 
     private String txid;
-    private BigDecimal vout;
+    private int vout;
     private BchTxScriptSig scriptSig;
     private String coinbase;
-    private BigDecimal sequence;
+    private long sequence;
 
 }
